Add new patches for the ready button
